+++ CHANGELOG.rst
@@ -13,6 +13,7 @@ Unreleased
 Added
 -----
 - Warning when trying to create rotations from large Euler angles
+- Passing symmetry when initializing an Orientation.
 - Vector3d.scatter() and Vector3d.draw_circle() methods to show unit vectors and
   great/small circles in stereographic projection
 - User guide with Jupyter notebooks as part of the Read the Docs documentation